requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# LLM & AI
openai>=1.0.0
anthropic>=0.18.0
langchain>=0.1.0
langchain-openai>=0.0.5
tiktoken>=0.5.0
faiss-cpu>=1.7.4
sentence-transformers>=2.3.0

# Korean Stock Market Data
finance-datareader>=0.9.31
pykrx>=1.0.40
yfinance>=0.2.35
psycopg2-binary>=2.9.9  # PostgreSQL 연결

# Financial Analysis
ta-lib>=0.4.28
empyrical>=0.5.5
pyfolio>=0.9.2

# Genetic Programming
deap>=1.4.1
gplearn>=0.4.2

# Data Processing
numba>=0.58.0
bottleneck>=1.3.7
pytables>=3.9.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.13.0
plotly>=5.18.0

# Web Interface
streamlit>=1.30.0
gradio>=4.0.0
fastapi>=0.109.0
uvicorn>=0.27.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
loguru>=0.7.0
tqdm>=4.66.0
joblib>=1.3.0
requests>=2.31.0  # KIS API 통신

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.5.0
